day_4/post_deployment_pipeline/model_server_new/requirements.txt
click==7.1.2
fastapi==0.65.1
h11==0.12.0
httptools==0.1.2
numpy==1.20.3
Pillow==8.2.0
prometheus-client==0.11.0
pydantic==1.8.2
pymongo==3.11.4
python-dotenv==0.17.1
python-multipart==0.0.5
PyYAML==5.4.1
six==1.16.0
starlette==0.14.2
torch==1.8.1
torchvision==0.9.1
typing-extensions==3.10.0.0
uvicorn==0.13.4
uvloop==0.15.2
watchgod==0.7
wheel==0.36.2
websockets==8.1
